split out sync map storage engine
